+++ public/demos/button-interactive/button-interactive-92edcfe2346b0560.js
@@ -327,22 +327,22 @@ function __wbg_get_imports() {
         }, arguments); },
         __wbindgen_generic_0000000000000001: function(arg0, arg1) {
             // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [NamedExternref("Event")], shim_idx: 6, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-            const ret = makeMutClosure(arg0, arg1, __wasm_bindgen_func_elem_79);
+            const ret = makeMutClosure(arg0, arg1, __wasm_bindgen_func_elem_78);
             return addHeapObject(ret);
         },
         __wbindgen_generic_0000000000000002: function(arg0, arg1) {
             // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [NamedExternref("KeyboardEvent")], shim_idx: 6, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-            const ret = makeMutClosure(arg0, arg1, __wasm_bindgen_func_elem_79_9);
+            const ret = makeMutClosure(arg0, arg1, __wasm_bindgen_func_elem_78_9);
             return addHeapObject(ret);
         },
         __wbindgen_generic_0000000000000003: function(arg0, arg1) {
             // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [NamedExternref("MouseEvent")], shim_idx: 6, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-            const ret = makeMutClosure(arg0, arg1, __wasm_bindgen_func_elem_79_10);
+            const ret = makeMutClosure(arg0, arg1, __wasm_bindgen_func_elem_78_10);
             return addHeapObject(ret);
         },
         __wbindgen_generic_0000000000000004: function(arg0, arg1) {
             // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [], shim_idx: 8, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-            const ret = makeMutClosure(arg0, arg1, __wasm_bindgen_func_elem_82);
+            const ret = makeMutClosure(arg0, arg1, __wasm_bindgen_func_elem_81);
             return addHeapObject(ret);
         },
         __wbindgen_generic_0000000000000005: function(arg0) {
@@ -369,20 +369,20 @@ function __wbg_get_imports() {
     };
 }
 
-function __wasm_bindgen_func_elem_82(arg0, arg1) {
-    wasm.__wasm_bindgen_func_elem_82(arg0, arg1);
+function __wasm_bindgen_func_elem_81(arg0, arg1) {
+    wasm.__wasm_bindgen_func_elem_81(arg0, arg1);
 }
 
-function __wasm_bindgen_func_elem_79(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_79(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_78(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_78(arg0, arg1, addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_79_9(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_79_9(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_78_9(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_78_9(arg0, arg1, addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_79_10(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_79_10(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_78_10(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_78_10(arg0, arg1, addHeapObject(arg2));
 }
 
 const TerminalDebugApiFinalization = (typeof FinalizationRegistry === 'undefined')